refactor(cli): let commands, targets and build steps declare themselves
Three central registries, each forcing unrelated features to edit the
same file: the command table, the `app build` target switch, and the
ensureAssets and runTest pipelines.

That is invisible until two features arrive at once — and then it is a
conflict in a file neither change is really about. It surfaced as an
ordering problem rather than a merge problem: seven branches with
exactly one real dependency between them could not be reviewed or
merged independently, because they all touched these files.

Now a thing is declared next to the code that implements it:

  register            a command
  registerTarget      a platform, and how to dispatch it
  registerAssetStep   work every build does, at a declared position
  registerPreTestStep provisioning some projects need and others do not

Measured: `app build web` went from editing five shared files to none.

Three things this changes rather than moves:

  - nounVerbs and the README table are derived from what registered.
    Both used to be written out by hand, which made them places to
    forget: a verb missing from the list dispatched correctly and
    appeared in no help at all.

  - buildTargets is a function. Go initialises package-level variables
    before it runs init, so a var reading the registry would have read
    an empty map and the CLI would have advertised no build targets.

  - printUsage and printCommandHelp derive their sections. Both
    hardcoded the noun list, so a noun a feature adds worked perfectly
    and appeared in no help at all.

Nothing may depend on init order — that is filename order, arbitrary,
and it changes when a file is renamed. Targets resolve after every init
has run, command order comes from an explicit field, and those fields
are spaced by ten so inserting one does not renumber the rest.

Tools are deliberately left alone. Their four scattered lists are the
same problem, and a change already exists that fixes it properly, with
mise integration so irgo stops downloading a second Node and JDK onto a
machine that already has both. That work is larger and better than a
narrower table added here would be.

Tests guard what decentralising costs: a lost registration is otherwise
silent, since everything derived from it agrees. TestTheCommandSetIsWhatWeThinkItIs
lists all twenty commands, and it caught `project pin` vanishing during
a rebase of this very change. TestNoGeneratedFilesAreTracked exists
because fifteen build artifacts were committed by a `git add -A` on a
branch whose .gitignore did not yet cover them.

Purely structural. `irgo --help`, all twenty command pages and every
noun page are byte-identical to before, which is how it was checked.
