+++ cmd/irgo/cmd_server.go
@@ -60,6 +60,43 @@ func runTest() error {
 	if err := ensureAssets(); err != nil {
 		return err
 	}
+	if err := runPreTestSteps(); err != nil {
+		return err
+	}
+
 	fmt.Println("Running tests...")
 	return runCommand(goBin(), "test", "./...")
+}
+
+func init() {
+	register(command{
+		noun: "project", verb: "test", order: 60,
+		summary: "Run the tests",
+		usage:   [][2]string{{"", "Regenerate assets, then go test ./..."}},
+		notes: `Assets first, so tests that render templates see the current ones rather than
+whatever was last on disk.`,
+	})
+}
+
+func init() {
+	register(command{
+		noun: "server", verb: "dev", order: 0,
+		summary: "Web server with hot reload",
+		usage:   [][2]string{{"", "Serve on :8080 and rebuild on save"}},
+		notes: `Rebuilds templ, Tailwind and the Go binary as you save. From an Android
+emulator the same server is http://10.0.2.2:8080, which is what
+irgo app run android --dev connects to.
+
+air is installed on demand.`,
+	})
+}
+
+func init() {
+	register(command{
+		noun: "server", verb: "serve", order: 10,
+		summary: "Web server without file watching",
+		usage:   [][2]string{{"", "Serve on :8080"}},
+		notes: `Regenerates assets once at startup and then leaves them alone — for checking a
+build, or running the web target.`,
+	})
 }
