+++ cmd/irgo/generated_test.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"os/exec"
+	"strings"
+	"testing"
+)
+
+// TestNoGeneratedFilesAreTracked — _templ.go, the Tailwind stylesheet and a
+// kit's copied assets are regenerated by every build. Committing them means
+// the repository carries a stale second copy that nothing refreshes, and a
+// diff full of noise nobody reads.
+//
+// This exists because it happened: fifteen of them were committed by a
+// `git add -A` run on a branch where the example's own .gitignore did not
+// exist yet, so nothing ignored them. They rode into every branch built on
+// that one. Nothing failed — generated files that are merely out of date do
+// not fail anything, which is exactly why a test has to say so.
+func TestNoGeneratedFilesAreTracked(t *testing.T) {
+	out, err := exec.Command("git", "ls-files").Output()
+	if err != nil {
+		t.Skip("not a git checkout")
+	}
+
+	// Suffix or path fragment that means "a build wrote this".
+	generated := []string{
+		"_templ.go",
+		"/static/css/output.css",
+		"/static/css/morpheus/",
+		"/static/js/morpheus.js",
+	}
+
+	for _, f := range strings.Split(string(out), "\n") {
+		f = strings.TrimSpace(f)
+		if f == "" {
+			continue
+		}
+		for _, g := range generated {
+			if strings.HasSuffix(f, g) || strings.Contains(f, g) {
+				t.Errorf("%s is tracked, but every build regenerates it — "+
+					"remove it with `git rm --cached` and let .gitignore hold it", f)
+			}
+		}
+	}
+}
